+++ Libplanet.Net/BoundPeer.cs
@@ -2,16 +2,14 @@ using System;
 using System.Diagnostics.Contracts;
 using System.Globalization;
 using System.Net;
-using System.Runtime.Serialization;
+using Bencodex;
 using Bencodex.Types;
 using Destructurama.Attributed;
 using Libplanet.Crypto;
-using Libplanet.Serialization;
 
 namespace Libplanet.Net
 {
-    [Serializable]
-    public sealed class BoundPeer : ISerializable, IEquatable<BoundPeer>
+    public sealed class BoundPeer : IEquatable<BoundPeer>, IBencodable
     {
         private static readonly byte[] PublicKeyKey = { 0x70 }; // 'p'
         private static readonly byte[] EndPointHostKey = { 0x68 }; // 'h'
@@ -32,18 +30,15 @@ namespace Libplanet.Net
         {
         }
 
-#pragma warning disable SA1118 // The parameter spans multiple lines
-        public BoundPeer(Bencodex.Types.Dictionary dictionary)
-            : this(
-                new PublicKey(((Binary)dictionary[PublicKeyKey]).ByteArray),
-                new DnsEndPoint(
-                    (Text)dictionary[EndPointHostKey], (Integer)dictionary[EndPointPortKey]),
-                dictionary[PublicIpAddressKey] is Text text
-                    ? IPAddress.Parse(text)
-                    : null)
+        public BoundPeer(Bencodex.Types.IValue bencoded)
+            : this(bencoded is Bencodex.Types.Dictionary dict
+                ? dict
+                : throw new ArgumentException(
+                    $"Given {nameof(bencoded)} must be of type " +
+                    $"{typeof(Bencodex.Types.Dictionary)}: {bencoded.GetType()}",
+                    nameof(bencoded)))
         {
         }
-#pragma warning restore SA1118
 
         internal BoundPeer(
             PublicKey publicKey,
@@ -62,16 +57,13 @@ namespace Libplanet.Net
             PublicIPAddress = publicIPAddress;
         }
 
-        private BoundPeer(SerializationInfo info, StreamingContext context)
+        private BoundPeer(Bencodex.Types.Dictionary bencoded)
+            : this(
+                new PublicKey(((Binary)bencoded[PublicKeyKey]).ByteArray),
+                new DnsEndPoint(
+                    (Text)bencoded[EndPointHostKey], (Integer)bencoded[EndPointPortKey]),
+                bencoded[PublicIpAddressKey] is Text text ? IPAddress.Parse(text) : null)
         {
-            PublicKey = new PublicKey(info.GetValue<byte[]>(nameof(PublicKey)));
-            EndPoint = new DnsEndPoint(
-                info.GetString("end_point_host"),
-                info.GetInt32("end_point_port"));
-            if (info.GetString(nameof(PublicIPAddress)) is string address)
-            {
-                PublicIPAddress = IPAddress.Parse(address);
-            }
         }
 
         /// <summary>
@@ -102,6 +94,17 @@ namespace Libplanet.Net
         public IPAddress? PublicIPAddress { get; }
 
         public string PeerString => $"{PublicKey},{EndPoint.Host},{EndPoint.Port}";
+
+        /// <inheritdoc/>
+        [Pure]
+        public Bencodex.Types.IValue Bencoded =>
+            Bencodex.Types.Dictionary.Empty
+                .Add(PublicKeyKey, PublicKey.Format(true))
+                .Add(EndPointHostKey, EndPoint.Host)
+                .Add(EndPointPortKey, EndPoint.Port)
+                .Add(
+                    PublicIpAddressKey,
+                    PublicIPAddress is IPAddress ip ? (IValue)(Text)ip.ToString() : Null.Value);
 
         public static bool operator ==(BoundPeer left, BoundPeer right) => left.Equals(right);
 
@@ -182,25 +185,6 @@ namespace Libplanet.Net
 
         public override int GetHashCode() => HashCode.Combine(
             HashCode.Combine(PublicKey.GetHashCode(), PublicIPAddress?.GetHashCode()), EndPoint);
-
-        public void GetObjectData(
-            SerializationInfo info,
-            StreamingContext context)
-        {
-            info.AddValue(nameof(PublicKey), PublicKey.Format(true));
-            info.AddValue("end_point_host", EndPoint.Host);
-            info.AddValue("end_point_port", EndPoint.Port);
-            info.AddValue(nameof(PublicIPAddress), PublicIPAddress?.ToString());
-        }
-
-        public Bencodex.Types.Dictionary ToBencodex() =>
-            Bencodex.Types.Dictionary.Empty
-                .Add(PublicKeyKey, PublicKey.Format(true))
-                .Add(EndPointHostKey, EndPoint.Host)
-                .Add(EndPointPortKey, EndPoint.Port)
-                .Add(
-                    PublicIpAddressKey,
-                    PublicIPAddress is IPAddress ip ? (IValue)(Text)ip.ToString() : Null.Value);
 
         /// <inheritdoc/>
         public override string ToString()