+++ Libplanet.Tests/Store/Trie/Nodes/ShortNodeTest.cs
@@ -1,4 +1,3 @@
-using System.Collections.Immutable;
 using Bencodex.Types;
 using Libplanet.Store.Trie.Nodes;
 using Xunit;
@@ -11,14 +10,14 @@ namespace Libplanet.Tests.Store.Trie.Nodes
         public void ToBencodex()
         {
             var shortNode = new ShortNode(
-                ByteUtil.ParseHex("beef").ToImmutableArray(),
+                ByteUtil.ParseHexToImmutable("beef"),
                 new ValueNode((Text)"foo")
             );
 
             var expected =
                 new List(new IValue[]
                 {
-                    (Binary)ByteUtil.ParseHex("beef"),
+                    (Binary)ByteUtil.ParseHexToImmutable("beef"),
                     new List(new IValue[] { Null.Value, (Text)"foo" }),
                 });
             var encoded = shortNode.ToBencodex();