+++ Libplanet/Blocks/Block.cs
@@ -2,8 +2,10 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Security.Cryptography;
+using System.Text.Json.Serialization;
 using Libplanet.Action;
 using Libplanet.Crypto;
+using Libplanet.JsonConverters;
 using Libplanet.Tx;
 
 namespace Libplanet.Blocks
@@ -87,6 +89,7 @@ namespace Libplanet.Blocks
         /// <summary>
         /// The <see cref="BlockHeader"/> of the block.
         /// </summary>
+        [JsonIgnore]
         public BlockHeader Header => _header;
 
         /// <inheritdoc cref="IBlockMetadata.ProtocolVersion"/>
@@ -96,9 +99,12 @@ namespace Libplanet.Blocks
         public BlockHash Hash => Header.Hash;
 
         /// <inheritdoc cref="IBlockHeader.Signature"/>
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+        [JsonConverter(typeof(ByteArrayJsonConverter))]
         public ImmutableArray<byte>? Signature => Header.Signature;
 
         /// <inheritdoc cref="IPreEvaluationBlockHeader.PreEvaluationHash"/>
+        [JsonConverter(typeof(ByteArrayJsonConverter))]
         public HashDigest<SHA256> PreEvaluationHash => _preEvaluationBlock.PreEvaluationHash;
 
         /// <inheritdoc cref="IBlockHeader.StateRootHash"/>