+++ Libplanet/Store/DefaultStore.cs
@@ -490,7 +490,7 @@ namespace Libplanet.Store
             UPath dirPath = path.GetDirectory();
             CreateDirectoryRecursively(_txExecutions, dirPath);
             using Stream f =
-                _txExecutions.OpenFile(path, System.IO.FileMode.OpenOrCreate, FileAccess.Write);
+                _txExecutions.OpenFile(path, System.IO.FileMode.Create, FileAccess.Write);
             Codec.Encode(SerializeTxExecution(txSuccess), f);
         }
 
@@ -501,7 +501,7 @@ namespace Libplanet.Store
             UPath dirPath = path.GetDirectory();
             CreateDirectoryRecursively(_txExecutions, dirPath);
             using Stream f =
-                _txExecutions.OpenFile(path, System.IO.FileMode.OpenOrCreate, FileAccess.Write);
+                _txExecutions.OpenFile(path, System.IO.FileMode.Create, FileAccess.Write);
             Codec.Encode(SerializeTxExecution(txFailure), f);
         }